+++ app/models/randomization_algorithm/randomization_algorithm.rb
@@ -1,13 +1,13 @@
 require 'randomization_algorithm/algorithms/default'
 require 'randomization_algorithm/algorithms/permuted_block'
-require 'randomization_algorithm/algorithms/minimisation'
+require 'randomization_algorithm/algorithms/minimization'
 
 module RandomizationAlgorithm
 
   DEFAULT_CLASS = RandomizationAlgorithm::Default
   ALGORITHM_CLASSES = {
     'permuted-block' => PermutedBlock,
-    'minimisation' => Minimisation
+    'minimization' => Minimization
   }
 
   def self.for(object)